Improve Log Request Content-Negotiation
The Content-Negotiation for getting logs of services has been improved
so that browsers that don't send `text/plain` in the accept header are
able to download the logs from the PREvant UI.
